--- react-design-concepts/src/3_Tabs Components/Tab.jsx
@@ -1,13 +1,0 @@
-import React from "react";
-
-const Tab = ({ label, isActive, onClick }) => {
-  return (
-    <div>
-      <div className={`tab ${isActive ? "active" : ""}`} onClick={onClick} >
-        {label}
-      </div>
-    </div>
-  );
-};
-
-export default Tab;
