+++ react-design-concepts/src/App.jsx
@@ -1,30 +1,14 @@
 import React from "react";
-// import MyComponents from './1_React_HOC/MyComponents'
-// import MyComponents from './2_Reder_Props_Pattern/MyComponents'
-// import MyComponents from "./2_Reder_Props_Pattern/Inputs";
-// import Main from "./3_Tabs Components/Main";
-// import Gallery from "./3_UseState/Gallery";
-import Performances from './4_Reack_Hooks/Performances'
+import Main from "./Tabs_Components/Main";
+import ProgressiveImage from "./Network_Optimize/Image";
 
 function App() {
   return (
     <div>
-      <Performances />
+      <Main />
+      <ProgressiveImage/>
     </div>
   );
 }
 
 export default App;
-
-
-/*
-1. React Components Pattern ? And its types to use?
-  - Functional Components
-  - Class Components
-  - Render Props
-  - Higher Order Component (HOC)
-  - React hooks 
-  - Tabs component  
-  Tabs are a common UI pattern used to organize content into multiple sections, 
-
-*/
