+++ react-design-concepts/src/Network_Optimize/3_Image.html
@@ -1,0 +1,58 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Image responsive With PIXCEL DPA</title>
+  </head>
+  <body
+    style="
+      max-width: 600px;
+      display: flex;
+      align-items: center;
+      justify-content: center;
+      min-height: 50vh;
+      flex-direction: column;
+    ">
+    <img
+      width="100%"
+      src="../assets/elephant_tiny.jpg"
+      srcset="
+        ../assets/elephant_tiny.jpg,
+        ../assets/elephant_lazy.jpg  2x,
+        ../assets/elephant_large.jpg 3x
+      " />
+
+    <img
+      width="100%"
+      src="../assets/elephant_tiny.jpg"
+      srcset="
+        ../assets/elephant_tiny.jpg   100w,
+        ../assets/elephant_lazy.jpg  500w,
+        ../assets/elephant_large.jpg 1000w
+      " />
+    <br />
+
+    <img
+      width="100%"
+      src="../assets/image.png"
+      srcset="
+        ../assets/image.png     100w,
+        ../assets/image@2x.png  300w,
+        ../assets/image@4x.png  600w,
+        ../assets/image.svg    1000w
+      " />
+    <br />
+    <img
+      width="100%"
+      src="../assets/image.png"
+      srcset="
+        ../assets/image.png,
+        ../assets/image@2x.png 2x,
+        ../assets/image@4x.png 3x,
+        ../assets/image.svg    4x
+      " />
+
+    x - means css pixel w - means device pixel
+  </body>
+</html>
