+++ react-design-concepts/src/Network_Optimize/Image.css
@@ -1,0 +1,12 @@
+body {
+  padding: 10px;
+}
+/* This is Loading */
+.loading {
+  filter: blur(10px);
+}
+/* This is Loaded */
+.loaded {
+  filter: blur(0px);
+  transition: filter 0.5s linear;
+}
