+++ react-design-concepts/src/Tabs_Components/Tabs.css
@@ -1,0 +1,51 @@
+* {
+  box-sizing: border-box;
+}
+.main {
+  width: 80%;
+  margin-inline: auto;
+}
+
+.tabs {
+  width: 100%;
+  display: flex;
+  gap: 10px;
+  flex-wrap: wrap;
+  align-items: center;
+  justify-content: space-evenly;
+}
+
+.tab {
+  flex-grow: 1;
+  font-family: sans-serif;
+  flex-basis: 40%;
+  cursor: pointer;
+  color: #333;
+  font-weight: 600;
+  font-size: 1rem;
+  text-transform: capitalize;
+  border: 1px solid #333;
+  padding: 5px 15px;
+  border-radius: 1vmax;
+  transition: 0.3s;
+}
+.tab:hover,
+.active {
+  color: #fff;
+  background-color: #333;
+}
+
+.content {
+  padding: 2rem;
+  margin-inline: auto;
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  gap: 1rem;
+  align-items: center;
+}
+@media (max-width:50em) {
+  .content{
+    display: block;
+    margin-inline: auto;
+  }
+}
